26-05-17 지역이미지나오게 지역컨트롤러 수정, 리뷰상세페이지 카페사진 가져오게 수정,26-05-17 무게별조회기능 수정 완벽구현 실패

+++ backend/src/main/java/com/ggori_salang/backend/vo/CafeVO.java
@@ -13,7 +13,7 @@ public class CafeVO {
     private String phone;
     private String description;
     private String allowedPetTypes;  // "소형견,중형견,고양이"
-    private double maxWeight;
+    private Integer maxWeight;
     private double rating;
     private int favoriteCount;
     private Double latitude;
